Refactor to combine map types into single component. Add handling for initial map render dimensions based on aspect ratio.
